+++ app/server/tests/main_test.go
@@ -3,12 +3,23 @@ package tests
 import (
 	"context"
 	"database/sql"
+	"errors"
 	"fmt"
 	"os"
 	"testing"
 
+	"github.com/golang-migrate/migrate/v4"
+	"github.com/golang-migrate/migrate/v4/database/postgres"
 	"github.com/joho/godotenv"
 )
+
+var db *sql.DB
+var ctx context.Context
+var err error
+
+type ResponseBody struct {
+	Message string `json:"message"`
+}
 
 func TestMain(m *testing.M) {
 
@@ -33,4 +44,34 @@ func TestMain(m *testing.M) {
 	code := m.Run()
 
 	os.Exit(code)
+}
+
+func runMigrations(db *sql.DB, order string) error {
+
+	driver, _ := postgres.WithInstance(db, &postgres.Config{})
+	m, err := migrate.NewWithDatabaseInstance(
+		"file://../../migrations/scripts",
+		"postgres", driver,
+	)
+
+	if err != nil {
+		return errors.New("Could not run migrations")
+	}
+
+	switch order {
+	case "up":
+		m.Up()
+	case "down":
+		m.Down()
+	default:
+		return errors.New("Migration run bad argument")
+	}
+
+	return nil
+}
+
+func failOnError(t *testing.T, err error) {
+	if err != nil {
+		t.Fatalf(err.Error())
+	}
 }
